Add metric logging
Can be turned on by setting LogMetrics to true for an agent before
starting it.

This just says to pretty-print out all metrics for all components for
every cycle. Unsuccessful sends will result in multiple writes for the
same metrics plus anything recorded since the last. This can ultimately
look a little weird, but it seemed better than dropping metric writes
entirely.

Pretty-printing of metrics is handled by just creating a tab writer so
that everything is nicely aligned per-component.